% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoEM-package.R
\docType{package}
\name{IsoEM-package}
\alias{IsoEM-package}
\alias{IsoEM}
\title{IsoEM: EM-Based Transcript Quantification for Long-Read RNA-seq}
\description{
IsoEM provides a modular pipeline for IsoQuant output quantification.
The workflow: \code{\link{prepare_isoem}} validates inputs;
\code{\link{build_ec}} / \code{\link{build_sc_ec}} construct equivalence
classes; \code{\link{run_em}} performs EM quantification;
\code{\link{write_isoem}} / \code{\link{write_sc_isoem}} write results.
One-step wrappers \code{\link{run_isoem}} and \code{\link{run_sc_isoem}}
are also provided.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hmutpw/IsoEM}
  \item Report bugs at \url{https://github.com/hmutpw/IsoEM/issues}
}

}
\author{
\strong{Maintainer}: hmutpw \email{hmutpw@github.com}

Authors:
\itemize{
  \item hmutpw \email{hmutpw@github.com}
}

}
