% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{print.IsoEMDataset}
\alias{print.IsoEMDataset}
\title{Print an IsoEMDataset object}
\usage{
\method{print}{IsoEMDataset}(x, ...)
}
\arguments{
\item{x}{An \code{IsoEMDataset} object.}

\item{...}{Ignored.}
}
\description{
Print an IsoEMDataset object
}
