% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{print.IsoEMEC}
\alias{print.IsoEMEC}
\title{Print an IsoEMEC object}
\usage{
\method{print}{IsoEMEC}(x, ...)
}
\arguments{
\item{x}{An \code{IsoEMEC} object.}

\item{...}{Ignored.}
}
\description{
Print an IsoEMEC object
}
