% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{print.IsoEMInput}
\alias{print.IsoEMInput}
\title{Print an IsoEMInput object}
\usage{
\method{print}{IsoEMInput}(x, ...)
}
\arguments{
\item{x}{An \code{IsoEMInput} object.}

\item{...}{Ignored.}
}
\description{
Print an IsoEMInput object
}
