% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{print.IsoEMResult}
\alias{print.IsoEMResult}
\title{Print an IsoEMResult object}
\usage{
\method{print}{IsoEMResult}(x, ...)
}
\arguments{
\item{x}{An \code{IsoEMResult} object.}

\item{...}{Ignored.}
}
\description{
Print an IsoEMResult object
}
