% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{print.IsoEMSCResult}
\alias{print.IsoEMSCResult}
\title{Print an IsoEMSCResult object}
\usage{
\method{print}{IsoEMSCResult}(x, ...)
}
\arguments{
\item{x}{An \code{IsoEMSCResult} object.}

\item{...}{Ignored.}
}
\description{
Print an IsoEMSCResult object
}
